# Refactor of beam search to process factor groups in parallel