@rhenry-nv 2021-03-11
This is broken as it does not handling forwarding hypotheses which could not be expanded by certain factor groups properly.